Final code for 2019-2020 Tower Takeover for State Comp 3/7/2020

+++ include/robot/subsystems/baseComponents/holonomicClass.hpp
@@ -12,10 +12,10 @@ private:
   Motor& m_backLeftMotor;
   Motor& m_backRightMotor;
 
-  int m_desiredFrontLeftVelocity = 0;
-  int m_desiredFrontRightVelocity = 0;
-  int m_desiredBackLeftVelocity = 0;
-  int m_desiredBackRightVelocity = 0;
+  double m_desiredFrontLeftVelocity = 0;
+  double m_desiredFrontRightVelocity = 0;
+  double m_desiredBackLeftVelocity = 0;
+  double m_desiredBackRightVelocity = 0;
 
   bool m_speedUpCurve = true;
 	bool m_driftTurning = true;
